pp3: tx queue select avoiding out-of-order
Objective:
Execution of Packet-stream TX from same application could
migrate form CPU to CPU. This causes for Out-Of-Ordering problem
(for example for iperf-udp application).

Solution has 2 points:
1). Implement and bind the ndo_select_queue() callback:
If (skb->ooo_okay) => use smp_processor_id (this is equal to previous)
else compute on an skb-hash base and set queue_mapping.
Skb-hash is permanent for an application and therefore qid is also
permanent and does not depend upon cpu.
2). USE the qid in driver instead of cpu-id

Change-Id: I193801f2ba53d304f659e04e6b23b8c4cf1051f4
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/37521
Tested-by: Star_Automation <star@marvell.com>
Reviewed-by: Dmitri Epshtein <dima@marvell.com>