pp3: allmulti for big mac-multicast list
Force/Auto switch dev-rx-mode to ALLMULTI
 if number of requested Multicast-MACs exceeds
 the max-capability MV_PP3_MAC_ADDR_NUM
There is a print about that, like:
 nic1: mac_mc(11) > max(10), force allmulti

Forced ALLMULTI is set in FW only but not in Linux-NET layer,
it is invisible in "ifconfig", could be seen by FW-sysfs command
   echo [PortNo] > /sys/devices/platform/pp3/fw/emac_vp_show

When Linux-NET changes mode it calls ndo_set_rx_mode() twice:
first with Current mode and then with New mode.
As result If mc-mac-list is already big
On transition NONE<->ALLMULTI or NONE<->PROMISC with big mc-mcast list
the print with "force allmulti" would be seen.

Change-Id: I45f9a80b4db477c58391cc86abcedd28c87d2437
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/44297